pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transformers-from-scratch"
version = "0.1.0"
description = "Transformers from Scratch"
requires-python = ">=3.11"
dependencies = [
  "aiohappyeyeballs==2.6.1",
  "aiohttp==3.13.2",
  "aiosignal==1.4.0",
  "anyio==4.12.0",
  "appnope==0.1.4",
  "asttokens==3.0.1",
  "attrs==25.4.0",
  "azure-identity>=1.25.3",
  "azure-mgmt-appcontainers>=4.0.0",
  "azure-mgmt-loganalytics>=13.1.1",
  "azure-mgmt-resource>=25.0.0",
  "azure-storage-blob>=12.28.0",
  "certifi==2025.11.12",
  "charset-normalizer==3.4.4",
  "comm==0.2.3",
  "contourpy==1.3.3",
  "cycler==0.12.1",
  "datasets==4.4.2",
  "debugpy==1.8.19",
  "decorator==5.2.1",
  "dill==0.4.0",
  "dotenv>=0.9.9",
  "executing==2.2.1",
  "filelock==3.20.1",
  "fonttools==4.61.1",
  "frozenlist==1.8.0",
  "fsspec==2025.10.0",
  "h11==0.16.0",
  "hf-xet==1.2.0",
  "httpcore==1.0.9",
  "httpx==0.28.1",
  "huggingface-hub==0.36.0",
  "idna==3.11",
  "ipykernel==7.1.0",
  "ipython==9.8.0",
  "ipython_pygments_lexers==1.1.1",
  "jedi==0.19.2",
  "Jinja2==3.1.6",
  "jupyter_client==8.7.0",
  "jupyter_core==5.9.1",
  "kiwisolver==1.4.9",
  "MarkupSafe==3.0.3",
  "matplotlib==3.10.8",
  "matplotlib-inline==0.2.1",
  "mpmath==1.3.0",
  "multidict==6.7.0",
  "multiprocess==0.70.18",
  "nest-asyncio==1.6.0",
  "networkx==3.6.1",
  "numpy==2.4.0",
  "packaging==25.0",
  "pandas==2.3.3",
  "parso==0.8.5",
  "pexpect==4.9.0",
  "pillow==12.0.0",
  "platformdirs==4.5.1",
  "prompt_toolkit==3.0.52",
  "propcache==0.4.1",
  "psutil==7.2.1",
  "ptyprocess==0.7.0",
  "pure_eval==0.2.3",
  "pyarrow==22.0.0",
  "Pygments==2.19.2",
  "pyparsing==3.3.1",
  "python-dateutil==2.9.0.post0",
  "pytz==2025.2",
  "PyYAML==6.0.3",
  "pyzmq==27.1.0",
  "regex==2025.11.3",
  "requests==2.32.5",
  "safetensors==0.7.0",
  "six==1.17.0",
  "stack-data==0.6.3",
  "sympy==1.14.0",
  "tiktoken==0.12.0",
  "tokenizers==0.22.1",
  "torch==2.9.1",
  "tornado==6.5.4",
  "tqdm==4.67.1",
  "traitlets==5.14.3",
  "transformers==4.57.3",
  "typing_extensions==4.15.0",
  "tzdata==2025.3",
  "urllib3==2.6.2",
  "wcwidth==0.2.14",
  "xxhash==3.6.0",
  "yarl==1.22.0",
]

[tool.setuptools]
packages = []

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'win32' or sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
